check_upstream_commits: Properly handle when provided SHAs were short
The logic that looks in the git log for SHA tags does not
properly handle the case where the regex includes text that
occurs after the SHA. This can cause a missed match when
the provided SHA was short. Update this case to consider
when extra hex characters are present after the initial
match is found.

Also update the abbreviated SHAs that are printed to be
the standard 12 character SHAs used by the Linux kernel.

Signed-off-by: Matthew R. Ochs <mochs@nvidia.com>
(cherry picked from commit fc493e85f8dadc7c854783566afb9cce0597ad89)